Implements graph structure
Creates a graph using an adjacency vector. I decided not to use a linked list to back the implementation
because I'm implmenting a pre-computation solution for shortest paths. This means that any change to the
graph structure will require re-construction, so there is no need to offer easy insertion and deletion of nodes
through a linked list backing structure. The array "vector" will work for my solution.