Implements more graph functionality and DirectPathExists check
Added functions to validate user input and to validate station "node" lookups.
This will all come in handy as more of the graph functionality is developed.
I need to spend time refactoring the interface of some functionality, and also
implement a better system to handle the off by one bug common when dealing with
1-based and 0-based indexing in different situations.